final_task/requirements.txt
beautifulsoup4==4.8.1
Jinja2==2.10.3
lxml==4.4.1
WeasyPrint==50
python-dateutil
